// docs/5.0/components/select-choices/select-choices.interface.ts
export interface SelectChoicesOption {
  label: string;
  value: string;
  disabled: boolean;
  cssClass?: string | string[];
  handler?: () => void;
}